+++ scripts/deploy.sh
@@ -3,15 +3,14 @@
 set -o errexit
 set -o pipefail
 set -o nounset
-
-echo $0
+shopt -s nullglob
 
 deps=${deps:-"False"}
 bootstrap=${bootstrap:-"False"}
 kill=${kill:-"False"}
 restart=${restart:-"False"}
 env=${env:-"True"}
-offline=${env:-"False"}
+offline=${offline:-"False"}
 
 echo "Building ubo-app"
 if [ "$offline" == "True" ]; then
@@ -26,15 +25,6 @@ if [ "$offline" == "True" ]; then
 else
   uv build --directory ubo_app/rpc
 fi
-
-for service in $(ls -d ubo_app/services/*/ubo-service); do
-  echo "Building service: $service"
-  if [ "$offline" == "True" ]; then
-    uv --offline build --directory "$service"
-  else
-    uv build --directory "$service"
-  fi
-done
 
 LATEST_UBO_APP_WHEEL=$(basename $(ls -rt dist/ubo_app-*.whl | tail -n 1))
 LATEST_BINDINGS_WHEEL=$(basename $(ls -rt dist/ubo_app_raw_bindings-*.whl | tail -n 1))
@@ -63,12 +53,9 @@ function run_on_pod_as_root() {
   return 1
 }
 
+run_on_pod_as_root "rm /tmp/ubo*.whl || true"
 scp dist/$LATEST_UBO_APP_WHEEL ubo-development-pod-$index:/tmp/
 scp dist/$LATEST_BINDINGS_WHEEL ubo-development-pod-$index:/tmp/
-for service in $(ls -d ubo_app/services/*/ubo-service); do
-  SERVICE_WHEEL=$(basename $(ls -rt "$service"/dist/*.whl | tail -n 1))
-  scp "$service"/dist/"$SERVICE_WHEEL" ubo-development-pod-$index:/tmp/
-done
 
 run_on_pod "$(if [ "$deps" == "True" ]; then echo "pip install --upgrade /tmp/$LATEST_UBO_APP_WHEEL &&"; fi)
 pip install --no-index --upgrade --force-reinstall --no-deps /tmp/$LATEST_UBO_APP_WHEEL &&
@@ -91,3 +78,10 @@ if [ "$kill" == "True" ] || [ "$restart" == "True" ]; then
 $(if [ "$restart" == "True" ]; then echo "systemctl --user restart ubo-app.service &&"; fi)
 true"
 fi
+
+for service in ubo_app/services/*/ubo-service; do
+  args=(--index="$index")
+  [[ "$deps" == "True" ]] && args+=("--deps")
+  [[ "$offline" == "True" ]] && args+=("--offline")
+  uv run --directory "$service" poe deploy-to-device "${args[@]}"
+done